Move basic AI to propulsion
