+++ src/Propulsion.h
@@ -7,6 +7,7 @@
 #include "Camera.h"
 #include "json/JsonUtils.h"
 #include "scenegraph/Model.h"
+#include "DynamicBody.h"
 
 class Propulsion
 {
@@ -14,7 +15,7 @@ class Propulsion
 		// Inits:
 		Propulsion();
 		virtual ~Propulsion() {};
-		void Init( SceneGraph::Model *m, int tank_mass, double effectiveExVel, float ang_Thrust );
+		void Init( DynamicBody *b, SceneGraph::Model *m, int tank_mass, double effectiveExVel, float ang_Thrust );
 		// TODO: This is here because of lack of shared enum between ShipType and this
 		void SetLinThrust( int i, float t ) { m_linThrust[i] = t; }
 
@@ -27,6 +28,11 @@ class Propulsion
 		inline double GetThrustUp() const { return m_linThrust[THRUSTER_UP]; }
 		double GetThrustMin() const;
 		vector3d GetThrustMax(const vector3d &dir) const;
+
+		double GetAccelFwd() const { return GetThrustFwd() / m_dBody->GetMass(); }
+		double GetAccelRev() const { return GetThrustRev() / m_dBody->GetMass(); }
+		double GetAccelUp() const { return GetThrustUp() / m_dBody->GetMass(); }
+		double GetAccelMin() const { return GetThrustMin() / m_dBody->GetMass(); };
 
 		inline void SetThrusterState(int axis, double level) {
 			if (m_thrusterFuel <= 0.f) level = 0.0;
@@ -41,7 +47,7 @@ class Propulsion
 		inline void ClearAngThrusterState() { m_angThrusters = vector3d(0,0,0); }
 
 		inline vector3d GetActualLinThrust() const { return m_thrusters * GetThrustMax( m_thrusters ); }
-		inline vector3d GetActualAngThrust() const { return m_ang_thrust * m_angThrusters; }
+		inline vector3d GetActualAngThrust() const { return m_angThrust * m_angThrusters; }
 
 		// Fuel
 		enum FuelState { // <enum scope='Ship' name=ShipFuelStatus prefix=FUEL_ public>
@@ -56,14 +62,30 @@ class Propulsion
 		inline double GetFuelReserve() const { return m_reserveFuel; }
 		inline void SetFuelReserve(const double f) { m_reserveFuel = Clamp( f, 0.0, 1.0 ); }
 		float GetFuelUseRate();
-		double GetSpeedReachedWithFuel( double mass ) const;
+		// available delta-V given the ship's current fuel minus reserve according to the Tsiolkovsky equation
+		double GetSpeedReachedWithFuel() const;
 		inline float FuelTankMassLeft() { return m_fuelTankMass * m_thrusterFuel; }
-		// This is necessary to avoid savegamebumps:
+		/* TODO: This is necessary to avoid savegamebumps:
+		 * better if avoid it on ship
+		*/
 		inline void SetFuelTankMass( int fTank ) { m_fuelTankMass = fTank; }
 		void UpdateFuel(const float timeStep);
 		inline bool IsFuelStateChanged() { return m_FuelStateChange; }
 
 		void Render(Graphics::Renderer *r, const Camera *camera, const vector3d &viewCoords, const matrix4x4d &viewTransform);
+
+		// AI Propulsion
+		void AIModelCoordsMatchAngVel(const vector3d &desiredAngVel, double softness);
+		void AIModelCoordsMatchSpeedRelTo(const vector3d &v, const DynamicBody *other);
+		void AIAccelToModelRelativeVelocity(const vector3d &v);
+
+		bool AIMatchVel(const vector3d &vel);
+		bool AIChangeVelBy(const vector3d &diffvel);		// acts in obj space
+		vector3d AIChangeVelDir(const vector3d &diffvel);	// world space, maintain direction
+		void AIMatchAngVelObjSpace(const vector3d &angvel);
+		double AIFaceUpdir(const vector3d &updir, double av=0);
+		double AIFaceDirection(const vector3d &dir, double av=0);
+		vector3d AIGetLeadDir(const Body *target, const vector3d& targaccel, double projspeed);
 
 	protected:
 		virtual void SaveToJson(Json::Value &jsonObj, Space *space);
@@ -81,7 +103,7 @@ class Propulsion
 
 		int m_fuelTankMass;
 		float m_linThrust[ THRUSTER_MAX ]; // It was THRUSTER_MAX
-		float m_ang_thrust;
+		float m_angThrust;
 		double m_effectiveExhaustVelocity;
 		double m_thrusterFuel;	// remaining fuel 0.0-1.0
 		double m_reserveFuel;	// 0-1, fuel not to touch for the current AI program
@@ -90,6 +112,7 @@ class Propulsion
 		vector3d m_angThrusters;
 
 		double m_power_mul;
+		const DynamicBody *m_dBody;
 		SceneGraph::Model *m_smodel;
 };
